+++ src/accelerate/utils/deepspeed.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .optimizer import AcceleratedOptimizer
-from .state import is_apex_available, is_deepspeed_available
+from ..optimizer import AcceleratedOptimizer
+from ..state import is_apex_available, is_deepspeed_available
 
 
 if is_deepspeed_available():
